[Application] Add Application Management tool
This tool provides a way to list, install and uninstall Applications
using the D-Bus API that will be provided by Crosswalk when it is run
using the run-as-service switch.

Fow now, only listing installed applications is tested. When Crosswalk
implements the install/uninstall functionality, there should be little
to no changes to this tool. But the code is included so the aim of
this tool is made clear.