REGRESSION(86246ac): Roll back chromium-crosswalk to 4b75b5f.
Go 6 revisions back in chromium-crosswalk and point to the same commit we
were pointing to before rebasing Chromium from 31.0.1650.12 to 31.0.1650.57.

It is a bit weird that we have quite a few commits being unused in
chromium-crosswalk, but the M31 rebase should not have changed that anyway
(plus it broke the build due to missing gstreamer dependencies in the spec
file for Tizen).

BUG=https://crosswalk-project.org/jira/browse/XWALK-471