REGRESSION(7b16c7d): Fix the %install stage on Tizen.
Due to the changes introduced in the commit mentioned above, we need to
change the default value we pass to the BUILDDIR_NAME spec file variable:
`out' is always added to the build directory anyway, so if one tried to
build a Tizen package in a fresh build directory the installation stage
would fail, as files such as the "xwalk" binary are now in
BUILDDIR_NAME/out/Release/xwalk instead of BUILDDIR_NAME/Release/xwalk.

Additionally, this also makes the default cause (ie. no BUILDDIR_NAME
specified) more consistent with the rest of the world, as we ended up
building in out/out/Release instead of just out/Release.

This wasn't caught by any bots because they all had old xwalk binaries lying
around in BUILDDIR_NAME/Release, generated before 7b16c7d went it, and
yesterday new canaries were not created because the build was broken due to
something else.