[Android] Fix the test case about testExtensionBroadcast.
verifyResult() will be called from sub and top frame. When it was
called from sub frame, the value of messageReceivedInTop was null,
then set the title to "Fail". In java code, the value of title was
"Fail", so this test case failed.

The fix is to ensure both of the frames have received the broadcast
message, then set the title to "Pass".

BUG=https://crosswalk-project.org/jira/browse/XWALK-530