[Tizen] Load audio-session-manager dynamically in the Browser process
audio-session-manager is using an _attribute_((constructor)) to
initialize code and setup signal handlers at startup. So, linking
with this library is causing a sandbox violation by executing the
code in Renderer Process, since the Renderer Process and the
Browser Process are linked with the same libraries.

To prevent this issue, we load the audio-session-manager library
dynamically in the Browser process.

BUG=https://crosswalk-project.org/jira/browse/XWALK-517