REGRESSION(cc5d7e7): Make the default "gbs build" case work again.
Hopefully this is the last regression for the time being. We cannot pass
--generator-output=. to gyp because its make generator does not work as
expected in that case and writes "srcdir =" instead of "srcdir = ." in the
Makefile.

We now only pass --depth and --generator-output if the BUILDDIR_NAME macro
has been set. One can still pass "--define BUILDDIR_NAME
/home/abuild/rpmbuild/BUILD/crosswalk/src" and break things, but let's say
this is an extreme corner case.

BUG=https://crosswalk-project.org/jira/browse/XWALK-532