[Extensions][Tizen] Change Crosswalk Processes SMACK label.
In order to keep its privileges to a bare minimum, the
Crosswalk processes (BP, RP, EP) need to:

1- BrowserProcess should use "crosswalk" label;
2- ExtensionProcess should change its SMACK label from "crosswalk" to its web app id;
3- RenderProcess should change its SMACK label from "crosswalk" to "crosswalk::renderer"
4- All processes but the BrowserProcess should drop their CAP_MAC_ADMIN capability ASAP.

The first step is accomplished by giving the "crosswalk" execution label (SMACK64EXEC), or
domain as called on the manifest file, to Crosswalk binary. Since all other process are
forked from the same binary, they will all have the same starting label.
We should then change that label accordingly.
This will guarantee that both EP and RP not only run with a different SMACK label
from the runtime (crosswalk) but it will also them run with a label that provides access to all
the capabilities needed by their corresponding web app.

The RenderProcess, on the other hand, needs an even more restricted label
due to its own sandboxed nature. The manifest then will only add what is needed for this process
to communicate with other processes (EP and BP) and nothing else.

Other SMACK rules files are added so we enable the BP and EP to talk
to each other (and to RP, of course).
For last but not least, this commit also adds the needed rules to get BP working again.
All of these rules can be easily spotted on the manifest file as the "<smack request" and
"<smack permit" tags. These rules were gathered by dmesg debugging SMACK on Tizen 2.x,
by using a kernel built with CONFIG_AUDIT=y.

The fourth step drops the capability that enables a process to change its SMACK label or
to execute other MAC related commands.