[Extensions] Make instance destruction test stricter
In the test, we wanted to ensure that instances (at the time "Contexts")
were destroyed correctly. But we were creating more instances than we
were using, so the checks were less strict.

Now that we only create an instance if the extension is actually used,
we can make the ASSERTs stricter.