[Extensions] Remove --disable-loading-extensions-on-demand switch
The feature was added for Crosswalk 2. We kept the switch to easily
remove it in Crosswalk 2 in case of regressions, with the plan to remove
it from the next release.

Since the feature has been stable and no regressions related to it were
found in the Crosswalk 2 beta branch so far, we are removing the
switch now.