[Tools] Convert lint.py to Unix line endings
Python on Windows should be able to recognize Unix line ending, but it
makes a difference for Unix-based, because we set it as a script and the
DOS line ending messes up the interpretation of "#!/usr/bin/env python"
line.

This means that we should be able to call on Unix-based systems:

  $ tools/lint.py

instead of having to call python with the script as argument.