[Extensions] Make xesh depend only on what it uses
Currently xesh (XWalk extension test shell) depends on the entire
xwalk_runtime target. This is not needed, since the shell was designed
to use only Extension system.

One drawback of the current approach is that unrelated changes end up
forcing a recompilation of xesh, this have some impact when running
ninja without target parameters.

Instead of depending on the xwalk_runtime target, this patch makes xesh
depend directly on the base libraries and include the sources from
extensions.gypi.

One drawback of the new approach is that it leads to compiling some
sources twice. A future change that would fix this would be to make
extensions system into a static library, and make both xwalk_runtime and
xesh depend on it.