[Extensions] Move kXWalkExternalExtensionsPath to extension switches
So that both xwalk_runtime and xesh (the test shell for extensions) can
use it. This prepares to make xesh build do not depend on the rest of
the runtime.

I was avoiding moving this constant here because it was not necessary
for extension system. But given the multiple users, I consider extension
itself a fair place to store this constant.