[Extensions] Move kXWalkExternalExtensionsPath to extension switches
So that both xwalk_runtime and xesh (the test shell for extensions) can
use it.

I was avoiding moving this constant here because it was not necessary
for extension system. But I think now that's fair place to store this
constant.

This change also makes xesh not include files from the runtime.