[Extensions] Compile extensions as a static library
Since extensions doesn't depend on other parts of Crosswalk, we can
compile it as a static library. One advantage is that now both
xwalk_runtime and xesh depend on that library, avoiding multiple
compilation of some source files.

Unittests will be moved in the next patch. Browsertests depend on
xwalk_runtime, so they still are part of the xwalk_browsertest binary.