[Android][SysApps] Not sharing same device_capabilities_api.js
Tizen DeviceCapabilities should use C++ native helpers as much as
possible. Since there is not corresponding for Java, those changes will
break the Android impl.

For now, duplicate the JavaScript shim for Tizen and Android. It might be
shared again if Java helpers are implemented same as C++ in future.