[Tizen] Add CAP_MAC_ADMIN to XWalk binary.
XWalk on Tizen must be able to change its smack label.
For doing that it needs the CAP_MAC_ADMIN capability to be set to
its binary.

We do that because XWalk uses the same binary to spawn all of its
processes: the BrowserProcess (BP, the Runtime), the RenderProcess (RP)
and the ExtensionProcess (EP). For instance, after being forked,
RP will need to change its smack label to xwalk_renderer in order
to lower that process' privilegies. On the other hand, EP will change
its smack label to the web app id.

After doing that both EP and RP will drop their CAP_MAC_ADMIN
capabilities.