[Extensions] Remove uses of deprecated v8::String::New()
As announced in
http://permalink.gmane.org/gmane.comp.web.chromium.devel/47781, Chromium
will change to break when using deprecated V8 functions. Since the
non-deprecated versions are already available for us, we can start to
use them before we hit the breakage.

This patch changes our uses of v8::String::New() to the corresponding
new functions that take the v8::Isolate* parameter.