[Android] Fix app will crash at exit if runtime lib not found
It's because, the code is trying to get a boolean value from
method onKepUp in runtime. It will return null if runtime lib
not found.
For other class like String, return null is OK. But for boolean,
java will try to turn Boolean object null into boolean, which
causes NullPointerException.

BUG=https://crosswalk-project.org/jira/browse/XWALK-627