Implement runtime features support for Crosswalk.
This commit implements the first part to support runtime enabled
features in Crosswalk. It contains a set of scripts and an .in
file to declare the features and generate code to easily check
wether a feature is stable, experimental or used for testing and if
it's enabled or not.

The way to check in C++ if the feature is enabled is as follow :

if(XWalkRuntimeEnabledFeatures::runAsServiceEnabled) ...

This commit doesn't yet implement the support to trigger these through
the command line.