[Android] Popup correct message if runtime lib is for wrong arch
For android, if the native library does not match the
cpu arch of the target device, the library file won't be
installed to the device.
So if we found that the runtime lib package exist without
libxwalkcore.so, it can be considered the runtime lib is
for another cpu architecture.
For embedded mode, the dialog will just nofity user without
the button to download from PlayStore.

TODO(wang16): If using arm Crosswalk on x86, it will work
via libhoudini, but it takes much time to do binary translation
at application startup. We should detect such case and make
Toast to users.

BUG=https://crosswalk-project.org/jira/browse/XWALK-606