[Application][Linux] Use ObjectManagerAdaptor for '/installed' interface
ObjectManagerAdaptor was created based on InstalledApplicationsManager,
so we now can use it to clean up things a bit. The

One non-trivial change is that the object path now is being defined by
the Manager instead of the individual objects. This was needed so we can
put the finding code in the ObjectManagerAdaptor.