[Runtime] Move ownership of XWalkContentBrowserClient to Crosswalk object
This is mostly so we take control of BrowserMainParts, that has both
important hook points and currently own different subsystems.

Our browser test infrastructure was also changed to create a Crosswalk
object instead of a XWalkContentBrowserClient. This is necessary because
some pieces of the client (including main parts) will depend on
Crosswalk object.

Embedded in this commit is a small cleanup: removed unnecessary
forwarding declarations and reset() calls to scoped pointers in
XWalkMainDelegate.