[Runtime] Move RuntimeContext creation to Crosswalk object
This patch exposes to Crosswalk object the two stages of initialization
of the Browser Process and moves creation of RuntimeContext to it.

By looking at this patch, one could conclude that Crosswalk object might
end up becoming XWalkBrowserMainParts. This is a valid point, but the
reason to create a different object is as follow:

- Crosswalk object have the reverse relationship that XWalkBrowserMainParts
  have with XWalkContentBrowserClient.

- BrowserMainParts ends up being owned by the framework. While it's
  scope might fit our case, it is not as clear as the approach of having
  a separated object.

- A minor reason: it is easier to fill the new entity iteratively
  instead of moving all the XWalkBrowserMainParts content at the same
  time. Gives us opportunity to review some object ownership decisions.

It's possible that in the future XWalkBrowserMainParts gets reduced to
just forwarding the important hooks to Crosswalk object. Because the
first and second reasons, the separation still makes sense.