Implements runtime flags for Xwalk.
A new class is added to handle the runtime flags where each feature is
declared as "stable" or "experimental".

This class contains the getters XWalkRuntimeFeatures::isMyFeatureEnabled()
used to know if a feature is on or off from the C++ code.

Note that this class also handles whether a feature has been overriden
by the command line (--enable/--disable-myfeature) by querying the
CommandLine object.