[Applications][Tizen] Fix having to export the HOME variable for user 'app'
When the user changes to the 'app' user (e.g. using 'su - app') the
HOME environment variable is not set by default, this variable is used
for finding the address of the D-Bus session bus.

This is separated in a different header so it can be used by other tools,
see later commits.