[Applications][Linux] Add a tool to launch applications
This tool will launch an Application using its AppID (either via
parameter or via argv[0]).  It will talk to Crosswalk using the
'org.crosswalkproject.Running.Manager' and
org.crosswalkproject.Running.Application' to run and follow the
Application lifetime.

On Tizen, this tool will use the appcode infrastructure to get more
information about the launched application state.

We re-use the functionality of setting the 'HOME' environment
variable that was introduced in the previous commit.

And to avoid polluting the main binary with #ifdef's we put
the platform specific code (dealing with Tizen's appcore) into
a separated header.