[Applications][Tizen] Add a tool to add/remove an application to the App DB
Tizen has a custom way to register/unregister applications into its
application registry, some of the steps need to run as system
administrator, so we separate the steps that need super user access
into a small binary that will be called by the core Crosswalk.

A small amount of effort is made to at least test-compile this tool on
Linux Desktop, to make it easier to detect that some Chromium API has
changed.