[Runtime] Add XWalkRunner to the Browser Process
The role of this object if to own, setup and teardown all the subsystems
of Crosswalk in the Browser Process. It will serve as a central point
for those subsystems.

Currently, this role is being fulfilled by different objects in
Crosswalk, including XWalkBrowserMainParts, XWalkContentBrowserClient,
RuntimeRegistry, RuntimeContext. We'll step by step move things from
those objects to XWalkRunner.

Instead of using one of the Content API concepts as our central point,
the idea is to create a separated object with this clear role.

XWalkRunner is created by XWalkMainDelegate and destroyed by it. It's
lifetime is the execution of the program. It will be created only when
executing the Browser Process because CreateContentBrowserClient() is
called only in that process.

The factory function for construction and the virtual destructor are
preparation for allowing we have different implementations for different
platforms. The approach is the same as was applied in
XWalkBrowserMainParts.

The motivation for this changed was also described by Mikhail and
Alexander in
https://docs.google.com/a/intel.com/document/d/1N0VVlut7JCWFbAEwTQc0loN9_D2JlOXRJprQeCbXY2U