Application extensions support multiple WEB apps
Application extensions support multiple WEB apps: the corresponding application instance is found by the aliased render process host id.

The initialization of application extensions is moved to ApplicationSystem class which contains the needed context to init the extensions. The application lookup is also done there.

The rest of changes in the patch is related to a necessity for the application to obtain the render process host id even before its first "Runtime" is created.