[Extensions] Buildsystem refactoring
The main point of this patch is to move build units from "includes" to self
contained modules. The reasons are:

- Includes have their path relative to the file which includes the .gypi. So if
another component not located at the root of xwalk depends on extensions
(sysapps for example), it won't work.

- Make it easier to track dependencies. Instead of adding all the dependencies
inside xwalk.gyp, we isolate them to the component that requires it.

- Have separated test runners for extensions, making run tests faster.