[Runtime] Add OnRenderProcess{Created,Gone} to XWalkRunner
RenderProcessHost lifecycle is important for XWalkRunner because some of
the subsystems in Crosswalk are implemented in terms of Crosswalk
extensions or have Crosswalk extensions as part of the interface.

Every time a new Render Process (Host) is created (which in Crosswalk
should be equivalent to a new application running), we have a proper
hook to create the extensions for that application in XWalkRunner.