[Runtime] Add platform specific subclasses of XWalkRunner
This is a preparation for moving the extension creation to
XWalkRunner (and its subclasses). Eventually I expect to be able to nuke
the XWalkBrowserMainParts variants, so all "platform specific"
behavior (at least with regards to initialization) will be in
XWalkRunner subclasses.

An interesting point: for the same reasons XWalkRunner::GetInstance()
exists, the Android and Tizen variants also exist. I've added a note
reinforcing that they should be used with care, but it's best to have a
clear way to reach the top objects (and expose things when a platform
needs) instead of making ad-hoc solutions.