[Runtime] Save registered events information in ApplicationData if the Application is launched by unpacked path.
Crosswalk supports launching unpacked application without install
the application first, and this is the same with current browsertests.

To make these types of applications able to handle registered events,
a solution is to save the events information in ApplicationData when
executing the main document. Events for these applications will not be
saved in ApplicationStorage.