[Android] Enable multiple threads for rasterization
On Android, since impl side painting is enabled as default, the
rasterization is scheduled to a seperate threads. However, the default
number of threads for rasterization is 1.

This commit is to calculate the appropriate number of raster threads
according to the number of CPU cores. It could take advantage of the
multi-core hardware capability and bring performance boost in
rasterizing, according to the result of some experiments.