[Tizen] Add WidgetContainerView and TizenSystemIndicatorWidget
The system indicator is implemented as a widget because it needs to
receive events and may also be an overlay on top of the rest of the
content, regular views do not support this. We add it to the container,
that is a view that doesn't draw anything, just passes the bounds
information to the topview layout.