Change default application icon on Tizen.
A default icon is used when installing by Crosswalk runtime if no icon
is given in the manifest.json. Current default application icon is
different between Android and Tizen. The Android's icon is a white "W" on
red circle; whereas the Tizen's icon is pedestrian crossing on green
background.

This PR is to let Android and Tizen share the same default icon -
white 'W' on red circle.

BUG = https://crosswalk-project.org/jira/browse/XWALK-800