[Android] Add "--tool-path" option for make_apk_test.
Because the xwalk_app_template target will be moved out of the "out" folder
on buildbot/trybot when the feature of generating separate APK for ARM and x86 arch
is ready. So add more flexible option "--tool-path" to replace the "--build-dir"
and "--target" options for the make_apk_test.py.