[Packaging] Pass --no-keep-memory to ld by default.
When building Crosswalk (Chromium, in fact) on an IVI chroot, ld ends up
running out of memory.

Work around the issue by passing --no-keep-memory to ld so that it optimizes
for memory consumption over speed. I have not noticed any significant drop
in speed, though.

Do not add any checks for IVI as it does not hurt to have this on all
profiles.