[Packaging] Be compatible with multiple Tizen versions and profiles.
Prepare for the future and start detecting features and versions so that we
can support other Tizen versions and profiles besides Tizen 2.1 Mobile.

Checking for specific profiles is not very future-proof. We therefore try to
look for either specific features that may be enabled in the current profile
(such as Wayland support) or for a certain Tizen major version (the
differences betwen Tizen 2 and Tizen 3 are big enough that we need different
patches and build dependencies, for example).