[Packaging] Add a Mesa patch for Wayland when building for Tizen.
Import a patch from the Ozone-Wayland repository to prevent crashes when
using Mesa and Ozone caused by the way Chromium has GL headers in both
third_party/khronos and third_party/mesa and the wrong ones (with wrong
pointer sizes) may be picked up.

This patch should actually be used by anyone testing Chromium on Wayland,
but this would require us to have a Chromium Mesa fork of our own. For now,
then, we're going to apply it automatically only for Tizen and try to solve
the issue upstream. It is tracked by http://crbug.com/266310