[SysApps] Remove usage of XWalkRuntimeFeatures
This was the only dependency that sysapps/ had from runtime/. We keep
the usage of the XWalkRuntimeFeatures (which parses command line) in
runtime/ and add specific functions to disable each API.

The unit tests now don't refer to XWalkRuntimeFeatures. We now handle
"SysApps is disable" by not creating the SysAppsManager at all. Two
small tests were added to check if disabling the APIs individually have
some effect.

This patch is a first step to clean up the handling of extension
creation in the Android and Tizen variants. At the moment, creation of
the extensions in those ports is bypassing the SysAppsManager.