[Android] Implement the functions of application working from SD-Card.
  Crosswalk for Android port doesn't support the functions of application
moving to SD-card and working from SD-Card. This is because the AndroidManifest.xml
in xwalk app template doesn't support the application installation settings for
"preferExternal" or "Auto", and the corresponding intent actions should be registered
as response to the mounting and unmounting of SD-Card. This feature is to enable
the SD-Card related function supports.

BUG=https://crosswalk-project.org/jira/browse/XWALK-774