[Android] Fix the crash issue of loading app from app:// scheme.
The fixed places:
1. As the resource rquest URL was canonicalized to lowercase in
runtime core, the host name exacted from the URL may be inconsitent
with the app package name, if there is capital letter in the package
name. So convert package name to lowercase before comparing these
two strings.
2. The unpermitted resource request should not be processed. Once
such request is received, the request job should be marked as done,
and should returned from the process function at once. So add the
missing return to the related place.

BUG=https://crosswalk-project.org/jira/browse/XWALK-808