[Extensions] Fix self-defined JS object override system's object issue
When we define a new JS object in JS API file, if the name is the same
with some system object, the system object will be overrode by self-defined
object.
For example, if we add a new extension named "window.screen", it will
override all the original window.screen object.