[Extensions] Fix xesh prompt
The InputWatcher is being told that it can read from input before we had
a chance to print the output from previous. Because we printed the
prompt at that moment, we ended up printing the next prompt before the
results.

With this change, we print the prompt only after the previous command
output is printed.