Remove "Runtime" and "Dialog" extensions
Runtime extension was added as a test for our internal extension
infrastructure and is not being used neither by Android or Tizen. There
are other extensions that test this infra, so it's OK to remove.

Dialog was an experimental extension added when Crosswalk started that
served as example. When we changed to Aura backend in Linux, it stopped
working. Given that is unmaintained we are better off removing it.