[Application] Compare application IDs in a case insensitive way
When loading a resource from an `app://` URL, the application ID is used as
a hostname.  GURL converts that to a lowercase string, which makes it
impossible to use the URL hostname as a key to ApplicationDataMap if
Crosswalk is loading resources from an application that has been packaged in
WGT format.

Modify the ApplicationDataMap typedef to perform case insensitive search.